feat(collision): Collision flag button (read desc)
WARNING: This is not working for Editable_Poly geometry YET
This button will add flags to col box and mesh modifiers on all selected objects. If there is no mesh modifier on one of the selection, it will add it. It will make sure to not have multiple mesh modifier.